+++ src/main/java/com/yaac/model/Utility/SaveFile.java
@@ -17,11 +17,11 @@ public class SaveFile implements Serializable {
     private int currentScore;
     private int highScore;
     private int lives;
-    private int volume;
+    private float volume;
     private boolean[] unlockedEngines;
     private  boolean[] unlockedWeapons;
 
-    public SaveFile(int score, int gems, int speedLvl, int bulletSpeedLvl, int bulletDmgLvl, int bulletRatioLvl, boolean shield, boolean speed, int engine, int weapon, int checkpoint, int currentScore, int highScore, int lives, boolean[] unlockedEngines, boolean[] unlockedWeapons, int volume) {
+    public SaveFile(int score, int gems, int speedLvl, int bulletSpeedLvl, int bulletDmgLvl, int bulletRatioLvl, boolean shield, boolean speed, int engine, int weapon, int checkpoint, int currentScore, int highScore, int lives, boolean[] unlockedEngines, boolean[] unlockedWeapons, float volume) {
         this.score = score;
         this.gems = gems;
         this.speedLvl = speedLvl;
@@ -126,6 +126,6 @@ public class SaveFile implements Serializable {
     public boolean[] getUnlockedWeapons() {return unlockedWeapons;}
     public void setUnlockedEngine(int index) {this.unlockedEngines[index] = true;}
     public void setUnlockedWeapon(int index) {this.unlockedWeapons[index] = true;}
-    public void setVolume(int volume) {this.volume = volume;}
-    public int getVolume() {return this.volume;}
+    public void setVolume(float volume) {this.volume = volume;}
+    public float getVolume() {return this.volume;}
 }
